refactor(locale): split locale lookup modules
Move Locale lookup behavior into private sibling modules.
Keep public data, method, and schema surfaces unchanged.
